refactor: move packages to new locations, remove no longer needed packages completely
