--- pkg/downloader/models/fetch/fetch_test.go
@@ -1,71 +1,0 @@
-package fetch_test
-
-import (
-	"testing"
-
-	"github.com/handsomefox/redditdl/pkg/downloader/models/fetch"
-)
-
-func TestIsValidURL(t *testing.T) {
-	t.Parallel()
-	type args struct {
-		str string
-	}
-	tests := []struct {
-		name string
-		args args
-		want bool
-	}{
-		{
-			name: "http://",
-			args: args{
-				str: "http://",
-			},
-			want: false,
-		},
-		{
-			name: "google.com",
-			args: args{
-				str: "google.com",
-			},
-			want: false,
-		},
-		{
-			name: "google",
-			args: args{
-				str: "google",
-			},
-			want: false,
-		},
-		{
-			name: "www.google",
-			args: args{
-				str: "www.google",
-			},
-			want: false,
-		},
-		{
-			name: "http://google.com",
-			args: args{
-				str: "http://google.com",
-			},
-			want: true,
-		},
-		{
-			name: "https://google.com",
-			args: args{
-				str: "https://google.com",
-			},
-			want: true,
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			if got := fetch.IsValidURL(tt.args.str); got != tt.want {
-				t.Errorf("IsURL() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
